Temporarily revert code for freeing devices on bad fd in select.
This began producing buffer overflows out of the blue in the
regression tests.  More weirdly, it was retroactive - revisions after
my commit of 2014-08-24T11:27:11 that introduced this change started
failing in the same way, even though they had passed regressions
before. This may reveal an underlying bug in free_device() that is
phase-of-moon dependent - to be investigated.