Exit from gps_sock_mainloop() if connection to gpsd is lost
The current mainloop function does not evaluate the return value of
gps_read(). If the socket is closed (because gpsd is stopped) the
mainloop will consume 99% CPU constantly doing a recv() from the socket
which returns zero. gps_read already returns -1 if this condition is
encountered.

This patch checks the return value of gps_read() and if it is -1 the
mainloop is exited. This is also what happens in gps_shm_mainloop().