In SConstruct, 'strip' becomes 'nostrip' and the sense is inverted.
(Default inverts too, so rge build still defaults to stripping.)

This is the only change required to make a build with all options False
be minimal.  It's preparation for a 'minimal' option to make life slightly
easier for the Android and embedded guys.

Required that the scons line for Android be retconned.

All regression tests pass.