Fix RTCM2 detection in packet.c
Regression tests pass.

Author's notes:
        RTCM2 support seems to be practically broken, if the stream
frequently includes the 0x40 byte. This byte causes packet.c to
temporarily think it is an AT1 stream, and stop calling rtcm2_decode()
while it thinks so.

        I have an RTCM stream from station id 0, and this particular station ID
causes the 0x40 byte to appear in virtually every RTCM message, so for that
stream, I have to wait through hundreds of RTCM messages before it is
positively identified and RTCM decoding starts to work.

        Patch attached, works for me.