Revise SiRF driver to be get satellites-used info from the skyview sentence.
Before, that information was partly coming from the navigation-solution message
and being pieced together with related data from svinfo. This way of doing
things is more consistent.

While this change required a rebuild of four check files, the effects are minor.
Fixes don't change, but some generated pseudo-NMEA does and the error
modeler can sometimes compute error estimates in cases where the old
code could not because of a determinant-zero condition.