Add do { ... } while (0) around assignment in macro
Add the customary do { ... } while (0) around this macro which expands
to executable code (as opposed to an expression).
I believe this is not necessary when, as in this case, the code is a
single line.
But a) it costs nothing, b) it might become multi-line one day, and c)
it looks wrong to my C programmer eyes - alarm bells ring every time I
read it, and I don't want to blunt that generally healthy reaction.