Add identify_failing_build_options script to devtools
This script is a helper script to run through various combinations
of build options as enumerated in the script looking for broken
builds.  This has helped identify several bugs and some other general
cleanup in the code base.

to run:

> devtools/identify_failing_build_options.py

it will run for as long as you let it trying to build different build
configurations with minimal=on to find failing ones.  It prints out
hopefully useful status as well as stderr but discards stdout from the
builds.  Any failing build configurations are logged to
failed_build_configs.txt and can be easily checked later to see why they
failed and how to fix them.