Alas, we can't use ICANON.
For the moment, we do nothing with the textual argument to
gpsd_optimize_io(). We'd like to use it to set ICANON mode and let the
kernel do all the line buffering itself.  The problem with this is
that VTIME stops working in ICANON, so a binary packet coming up the
wire would result in no further input being seen.