Widen PPS offset field in the three montors (NMEA, uBlox, SiRF) that report it.
To accomplish this, I had to shorted "PPS offset:" to "PPS:"; documentation
has been adjusted accordingly

In the NMEA and uBlox cases this gets us the required 22 characters.
The SiRF monitor is four shy of that.

The OnCore monitor gets PPS offset in-band and already has fields
sised to report that.