Worst reason to remove an ugly kluge ever.
This change removes some macros which assured both source and object
compatibility with client API version 6.0 in spite of some internal
renames.

But it turns out the gps.h we shipped had a broken inclusion and would
have blown up any compile if it was used from /usr/include, anyway.
So nobody can have built with the changed names and we can remove
this cruft.