Factor PPS-related members of struct gps_device_t into a pps_thread_t structure.
No logic changes, though it looks like there are two because two guards
that would always have failed when the code was compiled with pps=off
are now conditioned out.  Also, this code is offset-preserving
so as not to break link-time compatibility of libgpsd.

(This is the subtler approach...)

All regression tests pass.