gpsd-report() -> gpsd_log()
This change is done so we can add a "log" hook to the pps_thread_t
structure (this is not done yet) and harmonize with the name of the
outer logging function. If that name had been left as gpsd_report()
there would have been scope for bad confusion with the report_hook
member.

Also, remove two stray duplicative printf calls from the NMEA2000 driver
(drivers shouldn't have printfs!) and fix one typo.

This is a step towards factoring out ntplib. For that to happen, the
PPS thread code needs to be decoupled from the core session structure.

No logic changes. Object compatibility preserved. All regression tests pass.