Attempt to support /dev/pps devices.
For these devices:

1. The hunt loop is disabled.

2. When emitting a PPS report, all other (non-/dev/pps) devices are
checked for in-band time.  If no time is found the report is not
emitted. If time is ound its seconds field is incremented by 1 and it
is used

Every PPS from a device resets its online time, which should prevent
/dev/pps devices from timing out too often.

Not yet tested with a /dev/pps device.  All regression tests oass.