Address Savannah bug #45270: serial driver does not work properly on pipes.
Simon Artott wrote:

When using a pipe with the serial driver (e.g. stdin from another
process) to provide NMEA data, the fd is set to blocking mode and
gpsd_poll() gets stuck in a loop calling get_packet() because read()
always waits for more data.

I have to remove the "Switch back to blocking I/O now that CLOCAL is
set" code from serial.c to get it to work. The driver should detect if
the fd is a pipe and keep it set to non-blocking.