Prevent a subtle bug in the SiRF driver that sometimes caused NTP problems.
Simon Hradecky <shradecky@nomissoft.com> writes:

I just verified a bug in the driver_sirf.c (version 3.15, I just discovered
3.16 was released a few days ago, didn't yet check this version), routine
sirf_time_offset:

Line 637:

double retval = NAN;

This sets the default offset to NAN, which later results in the real time in
ntp shared memory etc. to be set to 0 with a valid time stamp, further
causing the NTP daemon to suddenly jump to year 1970, if the SIRF last status
is not known to the driver.

I believe, this default offset would be correct to read:

double retval = 0;

The problem with the real time set to 0 as result of the NAN time offset in
the SIRF driver (while the clock time was correct) in ntp shared memory
segment occurs frequently (but not always) with GlobalSat SIRF modules, in
our case EB5531RE. This can happen at any time of operation, it is possible
that the gpsd starts out entirely correct with proper time stamps in all
fields and suddenly forces the NTP to 0 seconds since 1970 because the offset
becomes NAN.