OS/X build fix for RTCM2 driver.
OSX 10.5, when building for PPC, defines _BIG_ENDIAN as 1, which
satisfies the BSD ifdef, but doesn't have the expected value. The
result is an "endianness macros are not defined" error. This bug goes
back at least as far as 3.14, but I don't think it was ever reported
here.

The fix is quite simple - moving the Darwin __BYTE_ORDER setup ahead
of the BSD __BYTE_ORDER setup so that the Darwin definitions take
precedence. This should have no effect on a platform that doesn't
define the __DARWIN_xxx macros.

The actual fix is from ryandesign@macports.org, though I added a
comment line clarifying the reason that the order matters.