Removes tabs from SConstruct.
In general, using tabs for indentation in Python is risky, since the
quantity of indentation is semantically significant but tabs aren't
treated as equivalent to the corresponding number of spaces (whatever
that might be).  Also, this file mostly uses spaces, with just a few
places failing to conform to its style.

There were three places using tabs for other than normal indentation.
Two were for emitted C source, where one was a #define where it didn't
really matter, and the other was one of several lines where only the
one used a tab.  The third case was a long string constant with
backslash continuation, which has been changed to use implicit
continuation and implicit concatenation, which is normally the
preferred approach.  Note that this changes the <space><tab><spaces>
to a single <space>, which is probably what was intended, anyway, and
the amount of whitespace doesn't actually matter in this context.

TESTED:
Manually examined the content of 'tipwidget', and ran a quick build
and regression test under Linux.