Fixes build for OSX 10.4.
This is a modified version of ryandesign@macports.org's patch,
originally applied to 3.14 and brought forward to 3.15/3.16.  The
original patch attempted to handle earlier OSX versions as well, but
that was obviously untested, since the 10.3 version of the define is
actually _TIMESPEC_DECLARED, not _TIMESPEC_DEFINED (though MacPorts
doesn't support anything older than 10.4, anyway).  It turns out that
building for 10.3 is broken for other reasons, but this patch should
at least fix this particular problem for 10.3 (and earlier?) as well
as 10.4.

Aside from expanding the #ifndef, this adds a #define _STRUCT_TIMESPEC
inside the conditional, for consistency with other such usage, even
though it's unlikely that these particular headers will be included
multiple times.

TESTED:
Verified that building for OSX 10.4 fails without this patch, and
succeeds with it.  Also tested builds for 10.5 and 10.9 (both x86 and
PPC in the former case), as well as Linux and xxxBSD builds.  Ran
regression tests in all cases except the failing 10.4 case.