Fixes manual use of regress-driver with binary logs.
At present, running regress-driver manually on binary logs may fail,
even though it works when run from scons.  This is due to running in
an unsanitized environment when run manually.

The problem is that sed is being used to filter the daemon output, and
it's not really kosher to be using sed on binary data.  Whether the
"working" cases involve its working "correctly" or merely
"reproducibly" is unclear, but what is clear is that in can choke on
binary data with certain settings of the LANG environment variable (in
particular, "en_US.UTF-8" is known to fail on ac12_binary).  The fix
is merely to unset LANG in regress-driver.

Failures on binary data are also not very informative due to diff
simply reporting "binary files differ".  That can be fixed by adding
the -b option to diff, but the resulting output can contain characters
that screw up the terminal, so it's probably not a good default.

TESTED:
Ran the full "scons check" with this change, as well as a manual
"./regress-driver test/daemon/*.log", and observed success in both
cases.

Signed-off-by: Jon Schlueter <jschlueter@redhat.com>