Fixes echo nonportability in regress-driver.
Some implementations of the echo command (including the bash builtin
when invoked as '/bin/sh') don't implement the -n option, resulting in
a "-n Elapsed time:" line followed by the time on another line.  This
change eliminates the -n by constructing the elapsed time message as a
single line (since there's no fully portable way to suppress the
newline after an echo).

This also redirects the elapsed time message to stderr, for
consistency with the header messages.

TESTED:
Ran and observed correct output.

Signed-off-by: Jon Schlueter <jschlueter@redhat.com>