Suppresses bloated regression diff when test produces no output.
At present, if a daemon regression test simply dies without producing
any output, the console (or log) gets a huge useless diff between the
expected output and the null output.  This change checks for that case
and replaces the diff with a single line indicating that no output was
produced.  It does not affect the success/failure decision.

If a test produces some output before dying, the behavior is
unchanged.  Although it would be useful to have a more intelligent
diff in that case, that would be a considerably more complicated
change, and is also a much less common case.

TESTED:
Tested a success case, a normal failing case (by tweaking the .chk
file), and a fatal error case (by making "python2" not work).
Observed expected results (including return code) in all cases.

Signed-off-by: Jon Schlueter <jschlueter@redhat.com>