Adds missing failure messages to rtcm-regress and aivdm-regress.
Both rtcm-regress failures, and the aivdm idempotency failures, were
missing the "Test FAILED!" messages when failing.  This adds them.

This does not include a more complicated fix to arrange a nonzero exit
status in the failing cases.  These tests deserve being rewritten to
make better use of Python, anyway.

TESTED:
Exploited the OpenBSD mktemp failure to test the failing cases, as
well as testing the normal success cases on OSX.

Signed-off-by: Jon Schlueter <jschlueter@redhat.com>