Fixes rtcm-regress and aivdm-regress for OpenBSD.
Traditionally, mktemp required that the Xs in the template be at the
end.  Although most modern versions of mktemp have relaxed this
restriction (in spite of what their manpages may say), the OpenBSD
mktemp still has this restriction, and fails with the templates
provided in these tests.  Rearranging the templates fixes this, at the
expense of no longer having a pure ".chk" extension on the temp files.

TESTED:
Verified that these previously failing tests now pass on OpenBSD 5.6
(as well as OSX).

Signed-off-by: Jon Schlueter <jschlueter@redhat.com>