Eliminates spurious gpsd errors with gpsfake in TCP mode.
Depending on timing, in TCP mode, gpsd may try to reconnect to fake.py
between the time that the logfile is complete and the time that the
remove_device is issued.  With no listener, this results in a set of
three error messages from gpsd.  They don't affect the test outcome,
since they only go to stderr and the "real" output is complete, but
they're annoying to the user.

The likelihood of this happening increases with longer logfiles and
with larger WRITE_PAD values.  With a 30ms WRITE_PAD (the default on
OSX), most logfiles exhibit this on all platforms, though it's not
seen on Linux with the default zero WRITE_PAD.

The fix is simply to avoid closing the listener after accepting the
incoming connection.  It's not necessary to explicitly service the
listener further, since the OS listen queue is sufficient to accept
the additional connection.

Closing the listener at drain time is almost acceptable, but still
occasionally fails due to the fact that drain() is (necessarily)
called before remove_device().  Not explicitly closing it at all is
acceptable, since Python closes it as part of the object cleanup.

TESTED:
Ran regress-driver in TCP mode with a 30ms WRITE_PAD on all logfiles
on three versions of OSX, as well as Linux, FreeBSD, and OpenBSD, both
with and without the fix.  Observed spurious errors in all cases
without the fix, and no cases with the fix.  Also verified that no
dangling sockets were left at the end of the runs.