Adds -q (quiet) option to regress-driver.
This allows reducing the verbosity when running multiple single-case
instances of regress-driver (e.g., for parallelism).  It is not the
same as redirecting stderr, since the "Processing" messages also go to
stderr, and are useful in this case.

TESTED:
Ran regress, build, and view modes, with and without -q, and verified
expected output.

Signed-off-by: Jon Schlueter <jon.schlueter@gmail.com>