Parallelizes gps-regress and gps-makeregress.
This change turns each of these tests into a set of individual
single-case targets, so that they can be run in parallel.  It still
uses regress-driver, but with a single logfile for each instance.  It
uses the new -q option to reduce the verbosity when run this way.  The
"herald" using gpsfake -T is turned into a separate target that's a
dependency of the others, so that it always runs first.

Because consuming a large number of PTYs would be antisocial (not to
mention possibly not working), it uses UDP mode.  TCP mode would also
work, but is slightly slower (albeit negligibly except on FreeBSD).

Since the output format is deficient in this mode (including not
having the summary at the end), it runs the old way whenever multiple
jobs aren't enabled.  But the single-case tests are still generated,
to permit running single cases (or sets of cases) by name.

It might be more conservative not to parallelize gps-makeregress, but
one can get that effect simply by not enabling multiple jobs.

This change also removes an obsolete comment about clearing the SHM
segments before the run.

TESTED:
Ran gps-regress and gps-makeregress, in both serial and parallel
modes, on three versions of OSX, as well as Linux, FreeBSD, and
OpenBSD.