Adds test name to regress-driver success message in single -q mode.
When running parallelized daemon tests, the output includes many
identical success messages that don't indicate which particular test
completed.  This change causes the logfile names to be reported in
that case.  It only applies to a successful run with the -q option,
where exactly one test was run and it wasn't "not found".

TESTED:
Ran the complete set of regression tests in parallel mode, as well as
manual cases both with and without -q with both one and two tests
specified.

Signed-off-by: Jon Schlueter <jon.schlueter@gmail.com>