Makes gpsfake -T take -S into account.
This makes it possible to view the "slow" as well as the normal
WRITE_PAD values with gpsfake -T, without having to look at the source
code.  It adds a new GetDelay() function to fake.py, which returns the
proper delay value, taking into account both the possible environment
variable and the "slow" option.  This is now used for all hree uses of
WRITE_PAD.

Because the environment override is now part of the function, the
WRITE_PAD constant is no longer modified by the environment value, but
instead reflects the platform default.

TESTED:
Ran the full set of regression tests, with both default and acceptably
shortened WRITE_PAD.  Also verified that a zero value causes trouble
(OSX), and that adding -S to the zero value makes it work.

Signed-off-by: Jon Schlueter <jon.schlueter@gmail.com>