Fixes dbusflags for some platforms.
The simple '-ldbus-1' isn't adequate for platforms where the dbus
library isn't located in a "usual" place, and where additional include
paths and/or additional library paths may be needed.  Obtaining the
flags via pkg-config (as is already done in some other cases) takes
care of this.

It appears that dbusflags is used in enough places that the
questionable LIBPATH overrides no longer break dbus linking with this
fix.

TESTED:
Verified that build-all with dbus_export=yes builds on OSX, Linux,
FreeBSD, and OpenBSD.