Fixes building with ncurses on FreeBSD (#47349).
This makes the change suggested in the bug, where the "--libs" option
to pkg-config is replaced with "--libs-only-L --libs-only-l", in order
to suppress the "-rpath" flag that crashes scons.

TESTED:
Built "build-all check" on three versions of OSX, as well as Linux,
FreeBSD, OpenBSD, and NetBSD.  NetBSD still has unresolved issues with
both dbus and ncurses, requiring those options to be turned off, but
otherwise everything passes.