Disables ncurses when library not found.
When none of the entries in the elif chain is able to find a library
for [n]curses, this turns off the option (with a message), similar to
the way some other options are treated.

It also clarifies the CheckPKG messages by showing that the "check" is
via pkg-config.  Otherwise one may see (e.g. on OpenBSD) the confusing
scenario where "ncurses... no" is followed by a working build with
[n]curses.

TESTED:
Ran build-all on OSX, Linux, FreeBSD, OpenBSD, and NetBSD, verifying
that ncurses was enabled and built successfully on all but NetBSD (see
NetBSD, of course).  Also verified expected "Checking..."  messages.