Fixes install breakage from commit 79800e8.
If the list of .pc files had been a list rather than a tuple, the
change would have worked.  But since it was a tuple, reducing it to
one item without including the comma as a "tuple flag" turned the
parens into simple parens, causing the "for" to iterate over the
characters of the single string instead of the strings.

Removing the now-superfluous tuple and iteration fixes it.

TESTED:
Verified that the previously failing install on Ubuntu works with this
fix.

Signed-off-by: Jon Schlueter <jon.schlueter@gmail.com>