Fixes cleaning of Qt objects.
The recent rework of the 'clean' handling broke the cleaning of the Qt
objects, which was missed because Qt wasn't enabled in the testing.
This fixes it by moving the Qt object setup outside the "not cleaning
or helping" conditional, while suppressing the call on pkg_config in
that case.

It also creates the variables 'cleaning' and 'helping', so as not to
have to make multiple calls on GetOption() for those options.

TESTED:
Ran "scons -c" after building with Qt, and verified that all build
products were removed.

Signed-off-by: Jon Schlueter <jon.schlueter@gmail.com>