Fixes Qt5 build on OSX.
The workaround for the -rpath bug in scons breaks Qt5 linking on OSX
because pkg-config --libs-only-* doesn't include the needed -framework
option.  The fix is to make the hack optional, and only use it in the
ncurses/tinfo cases where the -rpath is observed.  The other uses of
pkg-config now just use the usual --libs flag.

This also adds a message when Qt support is disabled due to the
library's not being found (similar to the handling of some other
libraries).

It also gets rid of a silly use of lambdas for pkg_config().

TESTED:
Ran build-all, -h, and -c on OSX with no Qt, Qt4, Qt5, and shared=no,
and verified that builds were successful and properly cleaned.  Also
ran build-all, -h, and -c (with Qt absent) on Linux, FreeBSD, OpenBSD
and NetBSD, with correct results (including not crashing on FreeBSD
and ncurses getting disabled on NetBSD).

Signed-off-by: Jon Schlueter <jon.schlueter@gmail.com>